Reorganise code structure
