+++ android/app/src/debug/gen/com/example/literature/R.java
@@ -1,0 +1,7 @@
+/*___Generated_by_IDEA___*/
+
+package com.example.literature;
+
+/* This stub is only used by the IDE. It is NOT the R class actually packed into the APK */
+public final class R {
+}
